requirements.txt
streamlit
pytesseract
pillow
opencv-python-headless
numpy